Refactor the code and restructure the project
